+++ packages/config-tailwind/src/index.ts
@@ -1,2 +1,0 @@
-export { default as tailwindConfig } from './tailwind.config';
-export { createThemePlugin } from './theme.plugin';