+++ packages/react-ui/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/af-select';
+export { AFInput } from './lib/af-input';
 export { AFButton } from './lib/af-button';
 export { AFCard } from './lib/af-card';