+++ packages/ui-react/tailwind.config.ts
@@ -1,9 +1,0 @@
-import type { Config } from 'tailwindcss';
-import { createThemePlugin } from '@kirdes-org/config-tailwind';
-
-const config: Pick<Config, 'content' | 'plugins' | 'prefix' | 'theme'> = {
-  content: ['./src/**/*.tsx', 'src/App.tsx'],
-  plugins: [createThemePlugin()],
-};
-
-export default config;